ntohs returns an uint16_t, as per SUSv3, so it can't be -1

+++ src/usr.bin/systat/netcmds.c
@@ -195,7 +195,7 @@ selectport(long port, int onoff)
 {
 	struct pitem *p;
 
-	if (ntohs(port) == -1) {
+	if ((short)ntohs(port) == -1) {
 		if (ports == 0)
 			return (0);
 		free((char *)ports), ports = 0;
